5286 Move ApiGatewayWebSocketOutput to its own package
