+++ java/query/query-runner/src/main/java/sleeper/query/runner/tracker/WebSocketQueryStatusReportDestination.java
@@ -24,7 +24,7 @@ import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.output.ResultsOutputInfo;
 import sleeper.query.core.tracker.QueryStatusReportListener;
-import sleeper.query.runner.output.ApiGatewayWebSocketOutput;
+import sleeper.query.runner.websocket.ApiGatewayWebSocketOutput;
 
 import java.io.IOException;
 import java.util.HashMap;
